ReyinMovies/requirements.txt
pymysql
lxml
requests
FontTools
selenium
pyexecjs